+++ omega/mainframe/default_components.json
@@ -1749,95 +1749,21 @@
   {
     "名称": "第三方_by温柔_公会系统",
     "描述": "当前仅有公会排名还未完善 请不要使用 其余功能均可使用 如果有bug请在群里面@风也温柔",
-    "是否禁用": true,
+    "是否禁用": false,
     "版本": "0.0.1",
     "来源": "Built-In",
     "配置": {
-      "可设置公会权限的选择器": "@a[m=1]",
-      "设置公会权限触发词": "set",
-      "隐藏菜单触发词": "setop",
-      "禁止设置公会坐标": [
-        [
-          0,
-          0,
-          0
-        ]
-      ],
+      "是否需要自带领地":false,
       "一级保护指令": [
         "title @a[x=[x],z=[z],dx=[Dx],dz=[Dz],dy=300,y=-60] actionbar §b[领地范围内]\n§e公会名字:§f[公会名字]",
         "gamemode 2 @a[[领地范围内非会员],x=[x],z=[z],dx=[Dx],dz=[Dz],dy=300,y=-60,m=0]"
       ],
-      "保护延迟时间[秒]": 1,
-      "各种公会所需计分板": {
-        "权限计分板": "权限",
-        "购买公会计分板": "挖掘",
-        "银行存入计分板": "挖掘"
-      },
-      "各种提示词": {
-        "返回地皮时提示词": "欢迎回到 [公会名字] ",
-        "无地皮时提示": "[警告] 你没有地皮",
-        "权限不足时提示": "你所在公会权限不足开启此功能",
-        "公会菜单商品显示": "[i] 商品名:[商品] 价格:[价格]",
-        "商店购买失败提示": "[错误] 余额不足",
-        "邀请对方进入公会时提示": "[公会名字] 邀请你进入公会(y/n)",
-        "管理菜单显示": "§b[会长菜单]\n§a输入 §e0 §a§l邀请他人进入公会\n§a输入 §e1 §a§l设置他人权限\n§a输入 §e2 §a§l发布公告\n§a输入 §e3 §a§l设置二级保护区域\n§a输入 §e4 §a§l踢出公会成员§e5 §a§l删除公会§e6 §a§l查看公会信息§e7 §a§l删除成员",
-        "设置公会权限菜单模板": "[i] [公会名字]",
-        "暂无公会存在时提示词": "[错误] 抱歉现在还未有任何公会存在",
-        "显示个人信息模板": "[计分板名字] : [计分板分数] ",
-        "显示个人信息菜单开头": "[个人信息菜单]\n所属公会:[公会名字]\n公会权限:[公会权限]\n本人权限:[本人权限]\n",
-        "购买领地后提示": "[成功购买公会] 你的公会名字[公会名字]\n起点坐标:[起点坐标]\n终点坐标:[终点坐标]\n中心坐标:[中心坐标]",
-        "op隐藏菜单": "[op隐藏菜单]\n输入 0 查看指定公会信息 \n输入 1 修改指定公会",
-        "op隐藏菜单公会详细信息显示": "[公会名字]\n会长:[会长]\n成员:[成员]\n权限:[权限]\n中心坐标:[中心坐标]\n范围:[起始坐标]----[终点坐标]"
-      },
-      "显示个人信息所需计分板": {
-        "挖掘": "[贡献值]"
-      },
-      "公会初始等级": 1,
-      "最低公会等级可开启的功能": {
-        "buff加成": 4,
-        "二级保护": 4,
-        "传送点设置": 4,
-        "商店功能": 4,
-        "设置二级保护区域": 4
-      },
-      "最低权限可使用功能": {
-        "可以t出其他成员": 3,
-        "可以发布公告": 2,
-        "可以改变他人权限": 2,
-        "可以设置二级保护区": 3,
-        "可以进入二级保护区域": 2
-      },
       "二级保护指令": [
         "execute @a[scores={[权限计分板]=..[最低权限]},x=[x],y=[y],z=[z],dy=[Dy],dz=[Dz],dx=[Dx]] [x] [y] [z] tp @s ~-1~~-1",
         "title @a[x=[x],y=[y],z=[z],dy=[Dy],dz=[Dz],dx=[Dx]] actionbar 你处于二级区域中范围"
       ],
+      "保护延迟时间[秒]": 1,
       "公会价格": "20000",
-      "公会等级称呼": {
-        "1": "一级公会",
-        "2": "二级公会",
-        "3": "三级公会",
-        "4": "四级公会"
-      },
-      "公会菜单显示": {
-        "0": "输入[i] 购买公会",
-        "1": "输入[i] 返回公会",
-        "2": "输入[i] 商店系统",
-        "3": "输入[i] 会长菜单",
-        "4": "输入[i] 公会排行",
-        "5": "输入[i] 个人信息"
-      },
-      "成员等级称呼": {
-        "0": "成员",
-        "1": "会员",
-        "2": "管理员",
-        "3": "副会长",
-        "4": "会长"
-      },
-      "提示词": "公会系统",
-      "触发词": [
-        "公会"
-      ],
-      "购买领地时限制器": "@a[name=[player],scores={[购买计分板]=[价格]..},m=!2]",
       "公会保护范围": {
         "1": [
           150,
@@ -1856,12 +1782,13 @@
           500
         ]
       },
+      "公会初始等级": 1,
       "公会商店": {
         "0": {
+          "使用的货币": "挖掘",
+          "商品价格": "200",
           "商品名字": "a",
           "商品英文": "a",
-          "商品价格": "200",
-          "使用的货币": "挖掘",
           "购买时执行指令": [
             "give @a[name=[购买者]] stone ",
             "scoreboard players remove @a[name=\"[购买者]\"] [计分板] [购买价格]",
@@ -1870,14 +1797,109 @@
           "购买时检测指令": "testfor @a[name=[购买者],scores={[计分板]=[购买价格]..}]"
         },
         "1": {
-          "商品名字": " ",
-          "商品英文": " ",
+          "使用的货币": " ",
           "商品价格": " ",
-          "使用的货币": " "
+          "商品名字": " ",
+          "商品英文": " "
         }
-      }
+      },
+      "公会等级称呼": {
+        "1": "一级公会",
+        "2": "二级公会",
+        "3": "三级公会",
+        "4": "四级公会"
+      },
+      "是否允许公会可以kick":false,
+      "可设置公会权限的选择器": "@a[m=1]",
+      "各种公会所需计分板": {
+        "权限计分板": "权限",
+        "购买公会计分板": "挖掘",
+        "银行存入计分板": "挖掘",
+        "贡献计分板":"挖掘",
+        "贡献计分板显示名称":"贡献榜单",
+        "兑换贡献计分板":"bi",
+        "兑换贡献计分板名称":"金币"
+      },
+      
+      "各种提示词": {
+        "公会菜单显示":"公会菜单\n0.创建公会\n1.返回公会\n2.公会商店\n3.退出公会\n4.兑换贡献\n5.公会管理（仅宗主/副宗主可执行）\n6.个人信息",
+        "op隐藏菜单": "[op隐藏菜单]\n输入 0 查看指定公会信息 \n输入 1 修改指定公会",
+        "op隐藏菜单公会详细信息显示": "[公会名字]\n会长:[会长]\n成员:[成员]\n权限:[权限]\n中心坐标:[中心坐标]\n范围:[起始坐标]----[终点坐标]",
+        "公会菜单商品显示": "[i] 商品名:[商品] 价格:[价格]",
+        "商店购买失败提示": "[错误] 余额不足",
+        "无地皮时提示": "[警告] 你没有地皮",
+        "无公会提示词":"[警告] 你没有公会",
+        "显示个人信息模板": "[计分板名字] : [计分板分数] ",
+        "显示个人信息菜单开头": "[个人信息菜单]\n所属公会:[公会名字]\n公会权限:[公会权限]\n本人权限:[本人权限]\n",
+        "暂无公会存在时提示词": "[错误] 抱歉现在还未有任何公会存在",
+        "权限不足时提示": "你所在公会权限不足开启此功能",
+        "管理菜单显示": "§b[会长菜单]\n§a输入 §e0 §a§l邀请他人进入公会\n§a输入 §e1 §a§l设置他人权限\n输入 §e2 §a§l设置公会传送点（仅会长）\n§a输入 §e3 §a§l设置二级保护区域\n§a输入 §e4 §a§l踢出公会成员\n输入 §e5 §a§l删除公会\n输入 §e6 §a§l查看公会信息\n输入 §e7 §a§l删除成员\n输入 §e8 §a§l升级公会",
+        "设置公会权限菜单模板": "[i] [公会名字]",
+        "购买领地后提示": "[成功购买公会] 你的公会名字[公会名字]\n起点坐标:[起点坐标]\n终点坐标:[终点坐标]\n中心坐标:[中心坐标]",
+        "返回地皮时提示词": "欢迎回到 [公会名字] ",
+        "邀请对方进入公会时提示": "[公会名字] 邀请你进入公会(y/n)",
+        "购买公会成功提示词":"[购买成功] 成功购买公会",
+        "退出公会成功提示词":"§e§l[成功退出]你的原有公会",
+        "会长退出公会提示词":"抱歉会长不能退出公会",
+        "兑换贡献时提示词1":"[请输入兑换的金额]",
+        "兑换贡献值计分板积分不足提示词":"你金币数额不足\n无法兑换",
+        "兑换贡献值成功提示词":"[提示] 兑换成功",
+        "邀请对方加入公会失败提示词":"抱歉对方为会长级别或者特殊原因无法加入公会",
+        "有地皮时未站在地皮范围内设置传送点提示词":"[错误] 请站在公会内",
+        "传送坐标设置成功提示词":"[提示] 设置成功",
+        "提示公会当前权限":"[提示] 公会权限为:[公会权限]",
+        "提升工会权限失败提示词":"[错误] 余额不足或者不符合要求",
+        "公会权限已达到最高":"权限已达到最高",
+        "提升公会权限成功":"提升公会权限成功"
+  
+      },
+      "每次公会升级为下一级所需贡献值":{
+        "1":"2000",
+        "2":"3000",
+        "3":"4000"
+  
+      },
+      "成员等级称呼": {
+        "0": "成员",
+        "1": "会员",
+        "2": "管理员",
+        "3": "副会长",
+        "4": "会长"
+      },
+      "提示词": "公会系统",
+      "显示个人信息所需计分板": {
+        "挖掘": "[贡献值]"
+      },
+      "最低公会等级可开启的功能": {
+        "buff加成": 4,
+        "二级保护": 4,
+        "传送点设置": 4,
+        "商店功能": 4,
+        "设置二级保护区域": 4
+      },
+      "最低权限可使用功能": {
+        "可以t出其他成员": 3,
+        "可以发布公告": 2,
+        "可以改变他人权限": 2,
+        "可以设置二级保护区": 3,
+        "可以进入二级保护区域": 2
+      },
+      "禁止设置公会坐标": [
+        [
+          0,
+          0,
+          0
+        ]
+      ],
+      "触发词": [
+        "公会"
+      ],
+      "设置公会权限触发词": "set",
+      "购买领地时限制器": "@a[name=[player],scores={[购买计分板]=[价格]..},m=!2]",
+      "隐藏菜单触发词": "setop"
     }
-  },
+  }  
+  ,
   {
     "名称": "第三方_by温柔_优化版本雪球菜单组件",
     "描述": "优化的雪球菜单",
